Move App into /components and add snapshot test
