--- src/App.test.js
@@ -1,10 +1,0 @@
-/* eslint-disable jest/expect-expect */
-import React from "react";
-import ReactDOM from "react-dom";
-import App from "./App";
-
-test("renders without crashing", () => {
-  const div = document.createElement("div");
-  ReactDOM.render(<App />, div);
-  ReactDOM.unmountComponentAtNode(div);
-});
